Version 4.2
Fixed a problem with the generated SVG related to masking
Maze generation and solving algorithms now use graphs
Added some checks
Added `isIntegerInRange()`
Reworked the alert system to use classes
The maximum height of the displayed maze is now proportional to the screen height

+++ scripts/maze/maze.mjs
@@ -1,18 +1,8 @@
-/*
-Squares are indexed left-to-right, top-to-bottom, starting from 0.
-For example, a maze of width 5 and height 4 is indexed in the following manner:
- 0  1  2  3  4
- 5  6  7  8  9
-10 11 12 13 14
-15 16 17 18 19
-
-Walls are split into horizontal and vertical ones
-Horizontal walls are indexed by the square on their left
-Vertical walls are indexed by the square on their top
-*/
 import { $ } from "jquery.mjs";
-import { downloadFile, svgElement, validate } from "../common.mjs";
-import {} from "./maze_util.mjs";
+import { downloadFile, svgElement, validateInputs } from "../common.mjs";
+import { isIntegerInRange } from "../common_util.mjs";
+import { generateAlgorithmFromString, GenerateParams, solveAlgorithmFromString } from "./maze_util.mjs";
+import { SquareMaze, SquareMazeGenerateParams } from "./shapes/maze_square.mjs";
 // The width of a square in the resulting SVG
 const MAZE_SQUARE_WIDTH = 20;
 // The radius of the rounded corners
@@ -27,9 +17,15 @@ const FINAL_TIME_PRECISION = 2;
  * Returns the SVG template of a maze of given size.
  * @param width The width of the maze.
  * @param height The height of the maze.
- * @returns The JQuery object containing the SVG template.
+ * @returns The JQuery object containing the SVG template, or `null` if the given parameters are invalid.
  */
-function mazeSVGTemplate(width, height) {
+function squareMazeSVGTemplate(width, height) {
+    if (!Number.isInteger(width) || width <= 0) {
+        return null;
+    }
+    if (!Number.isInteger(height) || width <= 0) {
+        return null;
+    }
     return svgElement("svg")
         .attr({
         id: "maze-img",
@@ -37,7 +33,7 @@ function mazeSVGTemplate(width, height) {
         height: height * MAZE_SQUARE_WIDTH + IMAGE_MARGIN * 2,
         viewBox: `-${IMAGE_MARGIN} -${IMAGE_MARGIN} ${width * MAZE_SQUARE_WIDTH + IMAGE_MARGIN * 2} ${height * MAZE_SQUARE_WIDTH + IMAGE_MARGIN * 2}`
     })
-        .append(svgElement("mask").attr("id", "mask")
+        .append(svgElement("mask").attr("id", "maze-mask").attr("maskUnits", "userSpaceOnUse")
         .append(svgElement("path").attr({
         d: `M${CORNER_RADIUS},0 `
             + `h${width * MAZE_SQUARE_WIDTH - CORNER_RADIUS * 2} `
@@ -53,7 +49,7 @@ function mazeSVGTemplate(width, height) {
     })))
         .append(svgElement("g").attr({
         id: "cropped",
-        mask: "url(#mask)"
+        mask: "url(#maze-mask)"
     })
         .append(svgElement("g").attr("id", "maze-solution"))
         .append(svgElement("g").attr("id", "maze-endpoints"))
@@ -64,7 +60,7 @@ function mazeSVGTemplate(width, height) {
  * @param maze The maze.
  * @param $mazeSVG The SVG to draw onto. Expects a template.
  */
-function drawMaze(maze, $mazeSVG) {
+function drawSquareMaze(maze, $mazeSVG) {
     const $gWalls = $mazeSVG.find("g#maze-walls");
     // Draw horizontal walls
     for (const wall of maze.hWalls) {
@@ -101,13 +97,22 @@ function drawMaze(maze, $mazeSVG) {
     }));
 }
 /**
- * Draws the solution.
+ * Draws the solution. Does nothing if the given parameters aren't valid.
  * @param width The width of the maze.
- * @param solution The array of squares tracing out the solution.
- * @param squaresEndpoints The array containing the starting and ending squares.
+ * @param height The height of the maze.
+ * @param solution An array of squares tracing out the solution.
+ * @param endpoints An array containing the start and the destination.
  * @param $mazeSVG The SVG to draw onto. Expects a template with a maze drawn onto it.
  */
-function drawSolution(width, solution, squaresEndpoints, $mazeSVG) {
+function drawSquareMazeSolution(width, height, solution, endpoints, $mazeSVG) {
+    if (!Number.isInteger(width) || width <= 0)
+        return;
+    if (!Number.isInteger(height) || height <= 0)
+        return;
+    if (solution.some(vertex => !isIntegerInRange(vertex, 0, width * height)))
+        return;
+    if (endpoints.some(vertex => !isIntegerInRange(vertex, 0, width * height)))
+        return;
     const $gSolution = $mazeSVG.find("g#maze-solution").empty();
     const $gEndpoints = $mazeSVG.find("g#maze-endpoints").empty();
     /**
@@ -130,7 +135,7 @@ function drawSolution(width, solution, squaresEndpoints, $mazeSVG) {
         drawSquare(square, $gSolution);
     }
     // Draw endpoints
-    for (const square of squaresEndpoints) {
+    for (const square of endpoints) {
         drawSquare(square, $gEndpoints);
     }
 }
@@ -157,7 +162,7 @@ $(() => {
     // "Generate" pressed
     $("button#generate").on("click", function () {
         const $validationTargets = $("input#width, input#height");
-        if (validate($validationTargets)) {
+        if (validateInputs($validationTargets)) {
             // Show generator and solver status
             $("div#subsec-generator-status").show();
             $("div#subsec-solver-status").hide();
@@ -172,25 +177,42 @@ $(() => {
             // Generate maze
             const width = Number($("input#width").val());
             const height = Number($("input#height").val());
-            const algorithm = $("select#generate-algorithm").val();
+            const algorithm = generateAlgorithmFromString($("select#generate-algorithm").val());
+            // Validate
+            const params = SquareMazeGenerateParams.tryNew(width, height);
+            if (algorithm === null) {
+                console.error("Invalid algorithm");
+                return;
+            }
+            if (params === null) {
+                console.error("Invalid parameters");
+                return;
+            }
+            const options = GenerateParams.newSquare(algorithm, params);
             // Spawn worker
             const worker = new Worker("/scripts/maze/maze_worker_generate.mjs", { type: "module" });
-            worker.postMessage({ width, height, algorithm });
+            worker.postMessage(options.toObject());
             worker.addEventListener("message", e => {
                 switch (e.data.msg) {
                     // Complete
                     case "complete":
                         worker.terminate();
-                        maze = e.data.maze;
+                        // Store received maze
+                        const receivedMaze = SquareMaze.fromObject(e.data.maze);
+                        if (receivedMaze === null) {
+                            console.error(`Received unknown maze format from generation worker: ${e.data.maze}`);
+                            return;
+                        }
+                        maze = receivedMaze;
                         $("span#generate-time").text(`${(e.data.time / 1000).toFixed(FINAL_TIME_PRECISION)}s`);
                         const startTime = Date.now();
                         // Disable canceling
                         $("button#generate-cancel").hide();
                         $("button#generate-cancel").off("click", cancel);
                         // Initialize SVG
-                        $mazeSVG = mazeSVGTemplate(maze.width, maze.height);
+                        $mazeSVG = squareMazeSVGTemplate(maze.width, maze.height);
                         // Draw maze
-                        drawMaze(maze, $mazeSVG);
+                        drawSquareMaze(maze, $mazeSVG);
                         $("div#maze-img-container").empty();
                         $("div#maze-img-container").append($mazeSVG.clone());
                         // Report maze drawing status
@@ -209,12 +231,13 @@ $(() => {
                         break;
                     // Progress report
                     case "progress":
-                        $("span#generate-progress").text(`${e.data.progress}%`);
+                        $("span#generate-progress").text(e.data.progress);
                         $("span#generate-time").text(`${(e.data.time / 1000).toFixed(TIME_PRECISION)}s`);
                         break;
                     // Unknown message
                     default:
                         console.error("Received unknown message from generation worker:", e);
+                        return;
                 }
             });
             // Handle canceling
@@ -238,7 +261,7 @@ $(() => {
     // "Solve" pressed
     $("button#solve").on("click", function () {
         const $validationTargets = $("input#start-x, input#start-y, input#end-x, input#end-y");
-        if (validate($validationTargets)) {
+        if (validateInputs($validationTargets)) {
             // Show solver status
             $("div#subsec-solver-status").show();
             // Reset solution drawing status
@@ -254,10 +277,15 @@ $(() => {
             const endX = Number($("input#end-x").val());
             const endY = Number($("input#end-y").val());
             const end = (endY - 1) * maze.width + (endX - 1);
-            const algorithm = $("select#solve-algorithm").val();
+            const algorithm = solveAlgorithmFromString($("select#solve-algorithm").val());
+            // Validate
+            if (algorithm === null) {
+                console.error("Invalid algorithm");
+                return;
+            }
             // Spawn worker
             const worker = new Worker("/scripts/maze/maze_worker_solve.mjs", { type: "module" });
-            worker.postMessage({ maze, start, end, algorithm });
+            worker.postMessage({ maze: maze.toObject(), start, end, algorithm });
             worker.addEventListener("message", e => {
                 switch (e.data.msg) {
                     // Complete
@@ -269,7 +297,7 @@ $(() => {
                         $("button#solve-cancel").hide();
                         $("button#solve-cancel").off("click", cancel);
                         // Draw solution
-                        drawSolution(maze.width, e.data.solution, [start, end], $mazeSVG);
+                        drawSquareMazeSolution(maze.width, maze.height, e.data.solution, [start, end], $mazeSVG);
                         $("div#maze-img-container").empty();
                         $("div#maze-img-container").append($mazeSVG.clone());
                         // Report solution drawing status
@@ -291,12 +319,13 @@ $(() => {
                         break;
                     // Progress report
                     case "progress":
-                        $("span#solve-progress").text(`${e.data.progress}%`);
+                        $("span#solve-progress").text(e.data.progress);
                         $("span#solve-time").text(`${(e.data.time / 1000).toFixed(TIME_PRECISION)}s`);
                         break;
                     // Unknown message
                     default:
                         console.error("Received unknown message from worker:", e);
+                        return;
                 }
             });
             // Handle canceling
@@ -357,6 +386,7 @@ $(() => {
         }
         else {
             console.error("Maze not yet generated");
+            return;
         }
     });
 });
